refactor(pubsub): new file for SubscriptionSession
I am about to add functionality to this class and I believe it would be
much easier to test if the class was in its own file.
