+++ google/cloud/pubsub/CMakeLists.txt
@@ -44,6 +44,8 @@ add_library(
     internal/publisher_stub.h
     internal/subscriber_stub.cc
     internal/subscriber_stub.h
+    internal/subscription_session.cc
+    internal/subscription_session.h
     internal/user_agent_prefix.cc
     internal/user_agent_prefix.h
     message.cc
@@ -145,6 +147,7 @@ function (google_cloud_cpp_pubsub_client_define_tests)
         internal/default_ack_handler_impl_test.cc
         internal/emulator_overrides_test.cc
         internal/ordering_key_publisher_connection_test.cc
+        internal/subscription_session_test.cc
         internal/user_agent_prefix_test.cc
         message_test.cc
         publisher_connection_test.cc
