+++ google/cloud/pubsub/pubsub_client.bzl
@@ -27,6 +27,7 @@ pubsub_client_hdrs = [
     "internal/ordering_key_publisher_connection.h",
     "internal/publisher_stub.h",
     "internal/subscriber_stub.h",
+    "internal/subscription_session.h",
     "internal/user_agent_prefix.h",
     "message.h",
     "publisher.h",
@@ -53,6 +54,7 @@ pubsub_client_srcs = [
     "internal/ordering_key_publisher_connection.cc",
     "internal/publisher_stub.cc",
     "internal/subscriber_stub.cc",
+    "internal/subscription_session.cc",
     "internal/user_agent_prefix.cc",
     "message.cc",
     "publisher.cc",
