+++ google/cloud/pubsub/subscriber_connection_test.cc
@@ -51,7 +51,7 @@ TEST(SubscriberConnectionTest, Basic) {
               google::pubsub::v1::AcknowledgeRequest const& request) {
             EXPECT_EQ(subscription.FullName(), request.subscription());
             EXPECT_FALSE(request.ack_ids().empty());
-            for (auto& id : request.ack_ids()) {
+            for (auto const& id : request.ack_ids()) {
               EXPECT_EQ("test-ack-id-0", id);
             }
             return make_ready_future(Status{});
@@ -92,82 +92,6 @@ TEST(SubscriberConnectionTest, PullFailure) {
   auto handler = [&](Message const&, AckHandler const&) {};
   auto response = subscriber->Subscribe({subscription.FullName(), handler});
   EXPECT_EQ(expected, response.get());
-}
-
-/// @test Verify callbacks are scheduled in the background threads.
-TEST(SubscriberConnectionTest, ScheduleCallbacks) {
-  auto mock = std::make_shared<pubsub_testing::MockSubscriberStub>();
-  Subscription const subscription("test-project", "test-subscription");
-
-  std::mutex mu;
-  int count = 0;
-  EXPECT_CALL(*mock, AsyncPull(_, _, _))
-      .Times(AtLeast(1))
-      .WillRepeatedly([&](google::cloud::CompletionQueue&,
-                          std::unique_ptr<grpc::ClientContext>,
-                          google::pubsub::v1::PullRequest const& request) {
-        EXPECT_EQ(subscription.FullName(), request.subscription());
-        google::pubsub::v1::PullResponse response;
-        for (int i = 0; i != 2; ++i) {
-          auto& m = *response.add_received_messages();
-          std::lock_guard<std::mutex> lk(mu);
-          m.set_ack_id("test-ack-id-" + std::to_string(count));
-          m.mutable_message()->set_message_id("test-message-id-" +
-                                              std::to_string(count));
-          ++count;
-        }
-        return make_ready_future(make_status_or(response));
-      });
-
-  std::atomic<int> expected_ack_id{0};
-  EXPECT_CALL(*mock, AsyncAcknowledge(_, _, _))
-      .Times(AtLeast(1))
-      .WillRepeatedly(
-          [&](google::cloud::CompletionQueue&,
-              std::unique_ptr<grpc::ClientContext>,
-              google::pubsub::v1::AcknowledgeRequest const& request) {
-            EXPECT_EQ(subscription.FullName(), request.subscription());
-            for (auto const& a : request.ack_ids()) {
-              EXPECT_EQ("test-ack-id-" + std::to_string(expected_ack_id), a);
-              ++expected_ack_id;
-            }
-            return make_ready_future(Status{});
-          });
-
-  google::cloud::CompletionQueue cq;
-  auto subscriber = pubsub_internal::MakeSubscriberConnection(
-      mock, ConnectionOptions{grpc::InsecureChannelCredentials()}
-                .DisableBackgroundThreads(cq));
-
-  std::vector<std::thread> tasks;
-  std::generate_n(std::back_inserter(tasks), 4,
-                  [&] { return std::thread([&cq] { cq.Run(); }); });
-  std::set<std::thread::id> ids;
-  auto const main_id = std::this_thread::get_id();
-  std::transform(tasks.begin(), tasks.end(), std::inserter(ids, ids.end()),
-                 [](std::thread const& t) { return t.get_id(); });
-
-  std::atomic<int> expected_message_id{0};
-  auto handler = [&](Message const& m, AckHandler h) {
-    EXPECT_EQ("test-message-id-" + std::to_string(expected_message_id),
-              m.message_id());
-    auto pos = ids.find(std::this_thread::get_id());
-    EXPECT_NE(ids.end(), pos);
-    EXPECT_NE(main_id, std::this_thread::get_id());
-    std::move(h).ack();
-    ++expected_message_id;
-  };
-  auto response = subscriber->Subscribe({subscription.FullName(), handler});
-
-  while (expected_ack_id.load() < 100) {
-    auto s = response.wait_for(std::chrono::milliseconds(5));
-    if (s != std::future_status::timeout) break;
-  }
-  response.cancel();
-  EXPECT_STATUS_OK(response.get());
-
-  cq.Shutdown();
-  for (auto& t : tasks) t.join();
 }
 
 }  // namespace
